[chore] Refactor billing vs subscription
