+++ fixbackend/billing/service.py
@@ -13,25 +13,52 @@
 #  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 
-from typing import Annotated, List, Optional
+import logging
+from datetime import datetime, timedelta
+from typing import Annotated, List
+from typing import Optional
 
 from fastapi import Depends
+from fixcloudutils.util import utc
+from prometheus_client import Counter
 
 from fixbackend.auth.models import User
-from fixbackend.billing_information.models import (
+from fixbackend.billing.models import (
     PaymentMethod,
     PaymentMethods,
     WorkspacePaymentMethods,
 )
 from fixbackend.dependencies import FixDependency, ServiceNames
+from fixbackend.domain_events.events import BillingEntryCreated
 from fixbackend.domain_events.events import ProductTierChanged
 from fixbackend.domain_events.publisher import DomainEventPublisher
 from fixbackend.errors import NotAllowed
-from fixbackend.ids import ProductTier, UserId, WorkspaceId
-from fixbackend.subscription.models import BillingEntry, SubscriptionMethod
+from fixbackend.ids import ProductTier, BillingPeriod
+from fixbackend.ids import UserId, WorkspaceId
+from fixbackend.metering.metering_repository import MeteringRepository
+from fixbackend.subscription.models import SubscriptionMethod, BillingEntry
 from fixbackend.subscription.subscription_repository import SubscriptionRepository
+from fixbackend.utils import start_of_next_period
 from fixbackend.workspaces.models import Workspace
 from fixbackend.workspaces.repository import WorkspaceRepository
+
+log = logging.getLogger(__name__)
+AccountsCharged = Counter("billing_accounts_charged", "Accounts charged by security tier", ["product_tier"])
+
+
+def compute_billing_period_factor(
+    *,
+    billing_time: datetime,
+    last_charged: datetime,
+    period_value: timedelta,
+) -> float:
+    delta = billing_time - last_charged
+    # Default should be a single month. For shorter or longer periods, we use a fraction/factor of the month
+    period_lower_bound = period_value - period_value * 0.1667  # - 5 days in case of 30 days
+    period_upper_bound = period_value + period_value * 0.1667  # + 5 days in case of 30 days
+    is_full_period = period_lower_bound < delta < period_upper_bound
+    billing_factor = 1.0 if is_full_period else delta / period_value
+    return billing_factor
 
 
 class BillingEntryService:
@@ -39,11 +66,15 @@ class BillingEntryService:
         self,
         subscription_repository: SubscriptionRepository,
         workspace_repository: WorkspaceRepository,
+        metering_repository: MeteringRepository,
         domain_event_sender: DomainEventPublisher,
+        billing_period: BillingPeriod,
     ) -> None:
         self.subscription_repository = subscription_repository
         self.workspace_repository = workspace_repository
+        self.metering_repository = metering_repository
         self.domain_event_sender = domain_event_sender
+        self.billing_period = billing_period
 
     async def list_billing_info(self, workspace_id: WorkspaceId) -> List[BillingEntry]:
         billing_entries = [
@@ -151,6 +182,81 @@ class BillingEntryService:
             workspace = await update_product_tier(new_product_tier)
 
         return workspace
+
+    async def create_billing_entry(
+        self, subscription: SubscriptionMethod, now: Optional[datetime] = None
+    ) -> Optional[BillingEntry]:
+        if not subscription.active:
+            log.info(f"AWS Marketplace: subscription {subscription.id} is not active")
+            return None
+        try:
+            billing_time = now or utc()
+
+            next_charge = start_of_next_period(period=self.billing_period, current_time=billing_time, hour=9)
+            match self.billing_period:
+                case "month":
+                    billing_period_value = timedelta(days=30)
+                case "day":
+                    billing_period_value = timedelta(days=1)
+
+            last_charged = subscription.last_charge_timestamp or billing_time
+            kind = subscription.__class__.__name__
+
+            month_factor = compute_billing_period_factor(
+                billing_time=billing_time,
+                last_charged=last_charged,
+                period_value=billing_period_value,
+            )
+            workspaces = await self.workspace_repository.list_workspaces_by_subscription_id(subscription.id)
+
+            for workspace in workspaces:
+                # Get the summaries for the last period, with at least 100 resources collected and at least 3 collects
+                summaries = [
+                    summary
+                    async for summary in self.metering_repository.collect_summary(
+                        workspace.id,
+                        start=last_charged,
+                        end=billing_time,
+                        min_resources_collected=100,
+                        min_nr_of_collects=3,
+                    )
+                ]
+
+                tiers = [summary.product_tier for summary in summaries]
+                # highest recorded tier
+                product_tier = max(tiers, default=ProductTier.Free)
+                # We only count the number of accounts, no matter how many runs we had
+                usage = int(len(summaries) * month_factor)
+                on_payment_free_tier = product_tier.paid is False
+                if on_payment_free_tier or usage == 0:
+                    log.info(f"{kind}: subscription {subscription.id} has no usage")
+                    # move the charge timestamp tp
+                    await self.subscription_repository.update_charge_timestamp(
+                        subscription.id, billing_time, next_charge
+                    )
+                    return None
+
+                log.info(f"{kind}: subscription {subscription.id} collected {usage} times: {summaries}")
+                AccountsCharged.labels(product_tier=product_tier.value).inc(usage)
+                billing_entry = await self.subscription_repository.add_billing_entry(
+                    subscription.id,
+                    workspace.id,
+                    product_tier,
+                    usage,
+                    last_charged,
+                    billing_time,
+                    next_charge,
+                )
+                await self.domain_event_sender.publish(
+                    BillingEntryCreated(workspace.id, subscription.id, product_tier, usage)
+                )
+                return billing_entry
+            else:
+                log.info(f"{kind}: subscription {subscription.id} has no workspace")
+                return None
+        except Exception:
+            log.error("Could not create a billing entry", exc_info=True)
+            raise
 
 
 def get_billing_entry_service(fix_dependency: FixDependency) -> BillingEntryService:
